% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption_plot.R
\name{caption_plot}
\alias{caption_plot}
\title{Insert left & right captions on a plot}
\usage{
caption_plot(plot, left_caption = "Source:",
  right_caption = "Hartford Funds Data Science", fontsize = 12)
}
\arguments{
\item{plot}{The existing plot object the caption will be added to}

\item{left_caption}{Optional caption on the left hand side.  Recommend using for source description.}

\item{right_caption}{Optional caption on the right hand side.  Recommend using for company name.}

\item{fontsize}{The fontsize for the captions.}
}
\value{
A grobTree object.  Unlike ggplot objects, this will not automatically be drawn.
}
\description{
Take an existing plot and add a horizontal line break at the bottom,
with space for a caption on the left and right side underneath.
}
\examples{
my_plot <- ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Sepal.Width)) +
 ggplot2::geom_point() +
 Rtfordfunds::theme_hf()

caption_plot(my_plot, "Source: iris dataset", "Hartford Funds")
}
