% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_history.R
\name{create_history}
\alias{create_history}
\title{Combine activity & transactions into a single history table}
\usage{
create_history(activity, transactions, at_level = c("person", "pe"),
  activity_using_level = c("person", "pe"),
  transaction_using_level = c("person", "pe"))
}
\arguments{
\item{activity}{Data frame of activities}

\item{transactions}{Data frame of transactions}

\item{at_level}{What level do we want results at (person or pe)?}

\item{activity_using_level}{What level do we want use to group activities?}

\item{transaction_using_level}{What level do we want use to group transactions?}
}
\value{
The history data frame
}
\description{
This creates a "newsfeed" like table at either the person or PE level
}
\examples{
\dontrun{
history <- activity \%>\%
    create_history(transactions,
                   at_level = "person",
                   activity_using_level = "person",
                   transaction_using_level = "pe")
}
}
