% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indiv_date_period.R
\name{create_indiv_date_period}
\alias{create_indiv_date_period}
\title{Create records for every individual, date & period}
\usage{
create_indiv_date_period(individuals = dplyr::tibble(id = letters[1:10]),
  dates = dplyr::tibble(as_of = seq.Date(Sys.Date() - 361, Sys.Date() -
  1, 30)), periods = dplyr::tibble(from = c(seq(0, 180, 30), 365), to =
  c(seq(30, 180, 30), 365, 730), period = paste("period", from, to, sep =
  "_")))
}
\arguments{
\item{individuals}{Data frame with a single id column for each "individual" (person or pe)}

\item{dates}{Data frame with a single as_of column for when we want to see the data "as of"}

\item{periods}{Data frame of day windows relative to the as_of date}
}
\value{
A data frame with every combination of individual, date & period
}
\description{
Given data frames for individuals, dates and periods,
create a data frame with a record for every combination.
}
