% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_new_project.R
\name{create_new_project}
\alias{create_new_project}
\title{Create a new RStudio project with sub-directories}
\usage{
create_new_project(proj, path = "//AD1WAYFPS001/Analytics")
}
\arguments{
\item{proj}{The name of the new project}

\item{path}{The path of the directory where the project folder will be created.
Defaults to the Analytics directory.}
}
\value{
nothing
}
\description{
Create a new RStudio project with sub-directories
}
\examples{
\dontrun{
  create_new_project("my_test_project")
}
}
