% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_since_last.R
\name{days_since_last}
\alias{days_since_last}
\title{Get the number of days since last activity}
\usage{
days_since_last(indiv_date, history)
}
\arguments{
\item{indiv_date}{Data frame with a column for individual id's and a column of as-of dates}

\item{history}{Data frame generated using the create_history function}
}
\value{
A data frame with columns for each activity type noting the days since last occurance.  If there was no previous occurance, the value is NA.
}
\description{
Given a data frame of individuals and as-of dates and another
data frame with a history of activities, find the number of days
since the last occurance of each activity for each inidividual and as-of date.
}
\examples{
\dontrun{
days_since_last(indiv_date, history)
}
}
