% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distinct_funds_sold.R
\name{distinct_funds_sold}
\alias{distinct_funds_sold}
\title{Get the number of distinct funds sold by an individual during some period}
\usage{
distinct_funds_sold(indiv_date_period, transactions)
}
\arguments{
\item{indiv_date_period}{Data frame of individuals, as of dates and periods from create_indiv_date_period}

\item{transactions}{Data frame of transactions, from a data warehouse query}
}
\value{
A data frame with individual id, date, period and #of distinct funds sold
}
\description{
Given an individual, as-of date and period of days, determine the number
of distinct funds that were sold by that individual during the peroid
}
\examples{
\dontrun{
distinct_funds_sold(indiv_date_period, transactions)
}
}
