% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easy_dbconnect.R
\name{easy_dbConnect}
\alias{easy_dbConnect}
\title{Easily connect to a SQL server by name}
\usage{
easy_dbConnect(server_name)
}
\arguments{
\item{server_name}{Name of the server to connect to (ods, mfdw, Xpressfeed, Factset...)
See Rtfordfunds::sql_servers for a complete list.}
}
\value{
An odbc connection
}
\description{
Easily connect to a SQL server by name
}
\examples{
con <- easy_dbConnect("ods")
}
