% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hf_palette}
\alias{hf_palette}
\title{Brand color palette for Hartford Funds}
\format{A named character vector with 12 values:
\describe{
  \item{name}{The name of the color}
  \item{value}{The hexcode of the color}
}}
\source{
data-raw/create raw data.R
}
\usage{
hf_palette
}
\description{
Primary and secondary colors taken from the brand guide
}
\keyword{datasets}
