% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sql_servers}
\alias{sql_servers}
\title{Names, servers and databases of frequently used SQL servers}
\format{A data frame with 5 rows and 4 variables:
\describe{
  \item{name}{Shorthand name for the server}
  \item{server}{Path to the server}
  \item{database}{Name of the database on the server}
}}
\source{
data-raw/create raw data.R
}
\usage{
sql_servers
}
\description{
A dataset containing the names, servers and databases of frequently used SQL servers
}
\keyword{datasets}
