% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_activity.R
\name{summarise_activity}
\alias{summarise_activity}
\title{Summarise tables at various possible levels}
\usage{
summarise_activity(df, id_map = id_map, at_level = c("person", "pe"),
  using_level = c("person", "pe"), group_vars = c("date", "metric"),
  count_distinct = NULL)
}
\arguments{
\item{df}{The data frame with some value to be summarised}

\item{id_map}{A two-column data frame mapping dim_pers_id to dim_prod_entty_id}

\item{at_level}{What level do we want the final results at?}

\item{using_level}{What level do we want to use to summarise?}

\item{group_vars}{Other than person/pe, what other grouping variables do we need?}

\item{count_distinct}{Do we also want to count the distinct values of something?}
}
\value{
A data frame at the at-level & group_vars level, with the sum of a value (and maybe distinct count of some attribute)
}
\description{
Given a data frame such as activity/transaction/asset with person and/or pe ids,
summarise values to various possible levels.
}
\examples{
\dontrun{
#Create the id_map
id_map <- person \%>\%
  filter(!is.na(dim_prod_entty_id)) \%>\%
  select(dim_pers_id, dim_prod_entty_id)

#Count of activities at the person-date-type level
activity \%>\%
    rename(metric = type) \%>\%
    mutate(value = 1) \%>\%
    summarise_activity(id_map = id_map,
                       at_level = "person",
                       using_level = "person",
                       group_vars = c("date", "metric"))

#Sum transactions at the PE level, then redistribute back to the person level
transactions \%>\%
#pivot before feeding into function
pivot_longer(cols = c(grosssales, redemptions),
             names_to = "metric",
             values_to = "value") \%>\%
  filter(value > 0) \%>\%
  summarise_activity(id_map = id_map,
                     at_level = "person",
                     using_level = "pe")
}
}
