% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_history.R
\name{summarise_history}
\alias{summarise_history}
\title{Get summary metrics for individuals across various dates and periods}
\usage{
summarise_history(indiv_date_period, history, summary_func = sum,
  value_fill = 0, force_all_columns = FALSE)
}
\arguments{
\item{indiv_date_period}{Data frame of individuals, as of dates and periods from create_indiv_date_period}

\item{history}{Data frame of past activities and transactions from create_history}

\item{summary_func}{The function we want to use to summarise}

\item{value_fill}{How should any missing values be filled?}

\item{force_all_columns}{Do we want to ensure a column is created for every metric/period, even if no one has any values?}
}
\value{
A data frame with a row for every individual and as-of date, and a column for every metric/period
}
\description{
Given the "history" data frame (which is a newsfeed of past activities and transactions),
get a sum, distinct count or whatever of each type of activity which
occurred in the period window for that individual and as-of date.
}
