% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hf.R
\name{theme_hf}
\alias{theme_hf}
\title{Hartford Funds branded theme for ggplot}
\usage{
theme_hf(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{The base font size}

\item{base_family}{The base font family}
}
\value{
A theme
}
\description{
This theme borrows heavily from ggthemes::theme_fivethirtyeight()
but adds some additional tweaks and HF brand colors.
}
\examples{
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Length, Sepal.Width, color = Petal.Length)) +
ggplot2::geom_point() +
theme_hf() +
ggplot2::scale_color_gradient(low = Rtfordfunds::hf_palette["Domestic Blue"],
                              high = Rtfordfunds::hf_palette["Accent Orange"])
}
